refactor(Export): Remove duplicate code
