+++ src/assets/wise5/classroomMonitor/dataExport/strategies/DiscussionComponentDataExportStrategy.ts
@@ -1,8 +1,8 @@
 import { removeHTMLTags } from '../../../common/string/string';
-import { AbstractDataExportStrategy } from './AbstractDataExportStrategy';
 import { millisecondsToDateTime } from '../../../common/datetime/datetime';
+import { ComponentDataExportStrategy } from './ComponentExportStrategy';
 
-export class DiscussionComponentDataExportStrategy extends AbstractDataExportStrategy {
+export class DiscussionComponentDataExportStrategy extends ComponentDataExportStrategy {
   defaultDiscussionColumnNames = [
     '#',
     'Workgroup ID',
@@ -33,38 +33,18 @@ export class DiscussionComponentDataExportStrategy extends AbstractDataExportStr
     'Post Text'
   ];
 
-  constructor(private nodeId: string, private component: any) {
-    super();
+  protected generateComponentHeaderRow(component: any, columnNameToNumber: any): string[] {
+    const headerRow = [...this.defaultDiscussionColumnNames];
+    this.populateColumnNameMappings(headerRow, columnNameToNumber);
+    return headerRow;
   }
 
-  /**
-   * Generate an export for a specific Discussion component.
-   * TODO: Move these Discussion export functions to the DiscussionService.
-   * @param nodeId The node id.
-   * @param component The component content object.
-   */
-  export() {
-    this.controller.showDownloadingExportMessage();
-    const components = [{ nodeId: this.nodeId, componentId: this.component.id }];
-    this.dataExportService.retrieveStudentData(components, true, false, true).subscribe(() => {
-      const columnNames = [];
-      const columnNameToNumber = {};
-      let rows = [this.generateDiscussionComponentHeaderRow(columnNames, columnNameToNumber)];
-      rows = rows.concat(
-        this.generateDiscussionComponentWorkRows(
-          this.component,
-          columnNames,
-          columnNameToNumber,
-          this.nodeId
-        )
-      );
-      const fileName = super.generateExportFileName(this.nodeId, this.component.id, 'discussion');
-      this.controller.generateCSVFile(rows, fileName);
-      this.controller.hideDownloadingExportMessage();
-    });
-  }
-
-  private generateDiscussionComponentWorkRows(component, columnNames, columnNameToNumber, nodeId) {
+  protected generateComponentWorkRows(
+    component: any,
+    columnNames: string[],
+    columnNameToNumber: any,
+    nodeId: string
+  ) {
     const rows = [];
     const componentStates = this.teacherDataService.getComponentStatesByComponentId(component.id);
     const structuredPosts = this.getStructuredPosts(componentStates);
@@ -232,20 +212,7 @@ export class DiscussionComponentDataExportStrategy extends AbstractDataExportStr
     parentPost.replies.push(replyComponentState);
   }
 
-  private generateDiscussionComponentHeaderRow(columnNames, columnNameToNumber) {
-    this.populateDiscussionColumnNames(columnNames, columnNameToNumber);
-    const headerRow = [];
-    for (let columnName of columnNames) {
-      headerRow.push(columnName);
-    }
-    return headerRow;
-  }
-
-  private populateDiscussionColumnNames(columnNames, columnNameToNumber) {
-    for (let c = 0; c < this.defaultDiscussionColumnNames.length; c++) {
-      const defaultDiscussionColumnName = this.defaultDiscussionColumnNames[c];
-      columnNameToNumber[defaultDiscussionColumnName] = c;
-      columnNames.push(defaultDiscussionColumnName);
-    }
+  protected getComponentTypeWithUnderscore(): string {
+    return 'discussion';
   }
 }
